extract utlity functions
