+++ backend/corpora/uu_course_descriptions/uu_course_descriptions.py
@@ -1,9 +1,6 @@
 from datetime import datetime
 import os
 from django.conf import settings
-import re
-from django.utils.html import strip_tags
-from langdetect import detect
 from ianalyzer_readers.readers.xlsx import XLSXReader
 from typing import Mapping, Callable
 
@@ -11,7 +8,7 @@ from addcorpus.python_corpora.corpus import FieldDefinition, XLSXCorpusDefinitio
 from addcorpus.es_mappings import text_mapping, main_content_mapping, keyword_mapping, int_mapping
 from addcorpus.python_corpora.extract import CSV, Combined, Pass, Constant, Metadata
 from addcorpus.python_corpora.filters import MultipleChoiceFilter
-from addcorpus.serializers import LanguageField
+from uu_course_descriptions.utils import html_to_text, language_name, detect_language
 
 FACULTIES = {
     'BETA': 'Betawetenschappen',
@@ -66,25 +63,6 @@ def filter_label(label):
 
     return get_content_with_label
 
-def html_to_text(content):
-    html_replacements = [
-        (r'<style.*</style>', ''),
-        (r'&nbsp;', ' '),
-        (r'<li>', '<li>- '),
-        (r'</li>', '</li>\n'),
-        (r'<br />', '<br />\n'),
-        (r'</p>', '</p>\n'),
-        (r'_x000D_', ' '), # excel quirk
-    ]
-
-    for pattern, repl in html_replacements:
-        content = re.sub(pattern, repl, content, flags=re.DOTALL)
-
-    plain = strip_tags(content)
-
-    stripped_lines = '\n'.join(filter(None, map(str.strip, plain.splitlines())))
-    return stripped_lines.strip()
-
 def content_extractor(label):
     return Pass(
         Combined(
@@ -101,23 +79,6 @@ def all_content_extractor():
         content_extractor('INHOUD'),
         transform='\n'.join
     )
-
-def detect_language(content):
-    if len(content) < 50:
-        return
-
-    try:
-        detected = detect(content)
-        if detected == 'af':
-            # dutch is sometimes mistaken for afrikaans
-            # but we know afrikaans is never actually used in this corpus
-            return 'nl'
-        return detected
-    except:
-        pass
-
-def language_name(language_code):
-    return LanguageField().to_representation(language_code)
 
 
 def filter_teacher_roles(data):
